+++ src/main/java/com/github/pagehelper/IPage.java
@@ -22,39 +22,17 @@
  * THE SOFTWARE.
  */
 
-package com.github.pagehelper.dialect.helper;
-
-import com.github.pagehelper.Page;
-import org.apache.ibatis.cache.CacheKey;
-import org.apache.ibatis.mapping.BoundSql;
-import org.apache.ibatis.mapping.MappedStatement;
-
-import java.util.Map;
+package com.github.pagehelper;
 
 /**
  * @author liuzh
  */
-public class SqlServer2012Dialect extends SqlServerDialect {
+public interface IPage {
 
-    @Override
-    public Object processPageParameter(MappedStatement ms, Map<String, Object> paramMap, Page page, BoundSql boundSql, CacheKey pageKey) {
-        paramMap.put(PAGEPARAMETER_FIRST, page.getStartRow());
-        paramMap.put(PAGEPARAMETER_SECOND, page.getPageSize());
-        //处理pageKey
-        pageKey.update(page.getStartRow());
-        pageKey.update(page.getPageSize());
-        //处理参数配置
-        handleParameter(boundSql, ms);
-        return paramMap;
-    }
+    Integer getPageNum();
 
-    @Override
-    public String getPageSql(String sql, Page page, CacheKey pageKey) {
-        StringBuilder sqlBuilder = new StringBuilder(sql.length() + 64);
-        sqlBuilder.append(sql);
-        sqlBuilder.append(" OFFSET ? ROWS FETCH NEXT ? ROWS ONLY ");
-        pageKey.update(page.getPageSize());
-        return sqlBuilder.toString();
-    }
+    Integer getPageSize();
+
+    String getOrderBy();
 
 }